feat: Organize view controllers
